Introduced test to validate nested topic type
There was already a corresponding test for validating the return type of binding models targeting relationships, but not one for targeting nested topics.

It's unlikely most consumers will use nested topics with the `ReverseTopicMappingService`, if only because the data requirements are cumbersome, but so far as that remains supported, we'll want to make sure this is addressed by the tests.

This required introducing a new `InvalidNestedTopicListTypeTopicBindingModel` (quite the mouthful!), which is similar to the existing `InvalidRelationshipListTypeTopicBindingModel`.