Removed unnecessary condition from Nested Topic validation
In a previous commit, the `listType` was updated to ensure it wouldn't be null by default it to `typeof(object)` (016815ba). Given that, there's no need to explicitly check for `listType is not null` here. That said, even if we did want to check for this, it should have been higher up in the boolean logic chain.