Ensure implicit defaults are tested
The `AttributeCollection` extensions (e.g., `GetBoolean()`, `GetInteger()`) introduced implicit defaults as part of OnTopic 5.0.0, but not all of the tests were updated to evaluate those, even if they evaluated explicit defaults. This is easy to add in as part of existing tests of default values.

As part of this, I rearranged the order so that the implicit default is always evaluated first, followed by the explicit defaults. This serves no functional benefit, but helps keep the order consistent.