Corrected empty attribute test
The empty attribute test was failing because calling `SetValue()` with an empty value will translate that to `null`, for the same reason that `GetValue()` will translate empty values to `null` in return. The backdoor—and, thus, the reason for this test!—is for `AttributeRecord` instances added directly. So that's what we'll do!