Extend `MarkClean()` test to evaluate version with `version`
Previously, there was a `MarkClean()` test that validated that the `version` could be set, but it only evaluated `MarkClean(version)` not `MarkClean(key, version)`. This update extends that existing test to cover this condition as well. As part of this, I also verify that any clean attributes don't have their `LastModified` value updated.