Exclude pass-throughs from code coverage
The `InsertItem()` on `MemberInfoCollection<T>` is a _bit_ more than a pass-through, but the logic is pretty basic, mostly intended to customize the exception thrown by the base class. With new guards in the constructor, it's highly unlikely that this would ever actually get triggered in our expected use cases, and since this is an `internal` class anyway, it's not worth setting up a unit test for that edge case.