Remove unnecessary infinite loop detection
An infinite loop should no longer be possible—at least not without specific effort to create one—due to the fact that `Enforce()` now calls `Register()` itself, and thus should, at most, get called three times (with the third being to return 'true'). As such, we can remove this logic, as well as the `_setCounter` it relied upon.