Introduced test for setting value types to empty strings
Currently, the `TypeMemberInfoCollection` doesn't support setting value types to null or empty unless they're a string—even if they're otherwise nullable. Ideally, we'll fix this in a future release to provide more flexibility (see #43). Until then, this test confirms that `SetPropertyValue()` returns false in this scenario, and doesn't update the value.