Updated event tests to use `CachedTopicRepository`
This continues to test the underlying `ObservableTopicRepository`, just as the original versions did, but _additionally_ tests the event bubbling of the `CachedTopicRepository`. It is true that this conflates two issues in the tests, but as a) these will typically be used together, b) the event bubbling code is trivial, and c) the unit tests are otherwise identical, it's convenient to combine these into the same tests.