Bug Fix: Ensured `Delete(…, isRecursive)` default is set
This is a bit of a goofy design consideration of how optional parameters work. The interface definition doesn't enforce use of the same defaults, or even that the parameter should be optional, nor does the interfaces default get inherited by implementers. This can break the implied contract of the interface without any warnings.

In this case, calling `Delete(topic)` against the `ITopicRepository` is allowed, but calling `Delete(topic)` against e.g. `ObservableTopicRepository` is not.

Since `CachedTopicRepository` is the main entry point for most implementations, and it inherits its `Delete()` passthrough from `ObservableTopicRepository`, this breaks the `Delete()` call for anyone relying on that optional parameter.

(The optional parameters in C# are a bit of a hack, and actually get compiled out as required parameters. The default values instead get compiled in the _calling_ code. The functionality is handy, but can cause a lot of problems if you're not careful with this.)