Introduced test for retrieving content type from in-memory topic graph
When initializing a new topic repository, or otherwise importing a topic graph, there are times where the `ContentTypeDescriptor` exists in the in-memory topic graph being passed into `GetContentTypeDescriptor()`, but doesn't (yet) exist in the `GetContentTypes()` cache, which is initialized from the underlying persistence store. In those cases, the `GetContentTypes()` cache is updated using the data from the `topic` being relayed to the `GetContentType()` method, thus offering a fallback option. This unit test verifies that functionality works as expected.