Fixed logic in `IsExtendedAttributeMismatch()` test
The topic being evaluated did not have an `Id`, and was thus being treated as `IsNew`. A recent update treats attributes as `IsDirty` if the associated topic `IsNew` and, thus, a new addition that evaluates that mismatched attributes aren't returned as clean (8776bc46) wasn't being properly evaluated. And, in fact, those topics _will_ currently be returned as clean, so with this test fixed, that assertion fails!

That behavior is unintuitive and possibly a bug. That will be revisited as part of #44. Until then, though, this fixes the unit test to maintain parity with the _current_ behavior.