Extended test to evaluate both relationships and references
With the introduction of topic references, the `IncomingRelationships` collection can contain both relationships _and_ topic references. But the unit test for evaluating `IncomingRelationships` was _exclusively_ testing relationships. This extends that test to _also_ evaluate topic references.

(This is a bit confusing. But the reason is that a topic reference is, essentially, a 1:n relationship. And, thus, while a topic can only reference one other topic per key, multiple topics can reference a given topic with the same key. As such, the reciprocal reference is stored as a relationship, not a reference.)