Bug Fix: Concurrency issue with deleting references
The `Delete()` method loops through `References` collection and, while its doing that, attempts to delete those references from the same collection it's looping over. This causes a concurrency issue. This can be resolved by _copying_ those references to a list first, and then looping over that.

This was exposed by the extension of the related unit test to check topic references alongside relationships (f5e03ea1). The relationships already had this handling in place.