Removed legacy support for `DateTime` from `AddParameter()`
The corresponding `AddParameter()` overload that accepts a `DateTime` object now _assumes_ `DateTime2`, and so there's no longer a case where any client will call the `private` overload with a `sqlDbType` set to `DateTime`. This code can safely be removed.

In the unlikely event we want to use a `DateTime` column in the future, this can be reintroduced. As an internal helper method we aren't concerned about making this comprehensive to support all possible scenarios, just the scenarios currently required by the OnTopic SQL database schema.