Bug Fix: `TopicReferencesDataTable` should accept null values
The `AddRow()` method was correctly setup to allow `targetTopicId` to be nullable, and to handle that later on in the method. But it also contained a `Contract.Requires()` for `targetTopicId` which effectively prevented it from being null. That's a mistake, and limits our testing capabilities, as exposed by a recent unit test (6e064cba).