Introduced test to validate deleting existing attribute
If a previous attribute is deleted, the database is expected to return a `null` value for that record. If the topic already exists in the `referenceTopic` with a valid attribute for that `attributeKey`, then that attribute should then be deleted. This test confirms that is true.

(In fact, this test fails—but that's due to identifying an underlying bug. That bug will be fixed in a subsequent commit.)