Bug Fix: Permit nullable attribute values
Attributes can be `null` when returned from the database. This happens when a previous attribute was deleted; the new version will use a `null` value.

Previously, the `SetIndexedAttributes()` extension method handled this by skipping `null` values. That was totally fine when this was only used to create new topics. As of OnTopic 5.0.0, however, we allow passing in a `referenceTopic`, which permits updating existing topics. This is most notable in `ITopicRepository.Refresh()`, where the goal is explicitly to update existing topics with new values.

If the `null` values are skipped, that will retain the previous values in memory. Worse, those values could end up getting inadvertently persisted under certain circumstances, thus reintroducing them into the permanent record.

This fixes that bug by allowing `topic.Attributes.SetValue()` to be called with the `null` value, thus overwriting any existing values. This necessitated introducing a new `GetNullableString()` extension method to handle the `DBNull` to `null` conversion.