Updated `StubTopicRepository.Load()` to raise `TopicLoaded` event
The underlying `TopicRepository` doesn't raise the `TopicLoaded` event, as it does for other events. Instead, it relies on the derived topic repository to raise the event. That's because the `TopicLoaded` event should only be fired when new data is loaded from the database, and the `TopicRepository` doesn't know when that is.

Technically, the `StubTopicRepository` never loads from the database as it works off of a hard-coded, in-memory topic graph. But it's intended to _simulate_ an `ITopicRepository` that _does_. As such, it should raise the `TopicLoaded` event. This will allow tests using this double to validate the rest of the `TopicLoaded` apparatus (such as the `TopicLoaded` event itself, as well as the `OnTopicLoaded()` method for raising it).