Bug Fix: Remove special handling for `uniqueKey`
As of OnTopic 5.0.0, we should expect that the `uniqueKey` being passed in is, in fact, a fully-qualified unique key—i.e., that it starts with `root`. The current logic resulted in `Load("root")` incorrectly trying to lookup `root:root`. Sloppy! And unnecessary. Removed. (This is technically a bug, but only in the test apparatus, and thus not being tracked via a GitHub issue.)