Extended test to account for additional conditions
The `TopicRepository.IsExtendedAttributeMismatch()` helper method evaluates a number of conditions which weren't fully represented by the current logic. For instance, arbitrary values are never condidered mismatched, nor are records without an `isExtendedAttribute` record. This update adds those in, and additionally adds in an attribute (`IsHidden`) that should _not_ be stored as an extended attribute, but is marked as one.