Extended `HasSettableMethod()` test to account for additional conditions
The `HasSettableMethod()` test checks a number of conditions which weren't addressed in the original test. This includes methods with multiple parameters, and methods with non-settable parameter types. To support testing these, the `MethodBasedViewModel` needed to be extended to include these (invalid) methods.

While I was at it, I updated the calls to use `nameof()` to help validate that they were calling into actual members, and not e.g. failing because of an invalid name.