Introduced `HasGettableMethod()` test w/ attribute constraint
There was previously a test which evaluated using the `MemberDispatcher()` without an `Attribute` constraint, but this bypassed conditions which optionally evaluated only members which contained a specified attribute. To support testing these, the `MethodBasedViewModel` needed to be extended to include a method with an arbitrary annotation.

In practice, we don't have any production code that uses this constraint for methods, but since the constructor accepts an attribute constraint, it would be confusing not to support this condition as part of the `HasGettableMethod()` member, and thus it's in place to avoid confusion should we discover the need for this later.

(YAGNI?)