Removed unnecessary exit condition
Prior in the `SetMethodValue()` method, there is a call to `HasSettableMethod()` which will return `false` if the `value?.GetType()` is `null`. Since the `GetType()` of `null` is `null`, this will always return false. As such, there isn't a need for this condition. That said, the call to `value` won't know that without more involved annotations, nor is it particularly obvious that `HasSettableMethod()` would guarantee that `value` isn't `null`. Given that, I'm instead moving this condition to the member validation, which makes the logic easier to read and reason through, while accomplishing the same thing.