Ensured `null` conditions are checked
When calling `TopicReferenceCollection.SetValue()` with a `null` value, no `TopicReferenceRecord` is created unless an existing `TopicReferenceRecord` with the same `Key` already exists.

This is a bit of an internal optimization to prevent keeping track of `null` records unless they represent a change to the current state. But it also means that our unit tests which relied upon `SetValue()` to establish a `null` value in order to test null conditions wasn't working as expected, as those references were simply never added in the first place.

To mitigate this, we're explicitly adding a new `TopicReferenceRecord` with a `null` value in these cases, instead of going through `SetValue()`. In a real world implementation, we'd almost always prefer going through `SetValue()` over this approach, but for test purposes it meets our needs—and actually reduces the footprint of the test surface slightly.