Corrected test for evaluating incorrect `DateTime` value
The test for evaluating an incorrect `DateTime` value was, itself, incorrect. It was evaluating an incorrect `AttributeKey`, which is what the next unit test after it does. It should be evaluating the correct `AttributeKey`, but with an incorrect value—i.e., a value not compatible with `DateTime`. This is fixed.