Corrected equality checks for `Topic`
Quite a few unit tests used `ReferenceEquals()` to determine if the reference to two objects were equal. This is a valid approach, but not directly off of `Assert`—it would need to be wrapped in `Assert.IsTrue()`. This is because `ReferenceEquals()` is, of course, a method inherited from `object`, and makes no assertions that the unit test framework will recognize; instead, it simply returns `true` or `false`. Whooooops!