Exclude test model from code coverage
Test models are there to support unit tests, and needn't be the subject of testing themselves. In this case, the methods are there to validate whether or not they pass validation with e.g. `TypeMemberInfoCollection.HasSettableMethod()`, and aren't expected to actually be called.

Most test models don't have methods—and, in the rare case that they do, they're usually executed by the tests they were designed for. As a result, it's rare that we need to explicitly exclude test models from code coverage. The `MethodBasedViewModel` is an exception in that it, by design, uses methods instead of properties, and thus is perceived to have more logic than auto-implemented properties typically used.