Introduced test for evaluating mapped property conversion
If there's a source property with the same name as the target property, they will typically have the same data type—in which case they'll simply use the `SetCompatibleProperty()` method, instead of using `SetScalarValue()`. This test creates a contrived example where the source property type doesn't match the target property type, _but_ the `MemberDispatcher` is capable of converting between them (in this case, from an appropriately formatted string to a Boolean value). This is probably a rare case, but should be supported.

This required adding a `BooleanAsString` property to `CustomTopic`, as well as a new `ConvertPropertyViewModel`, which a `bool` that can convert from string, and a `string` that can convert from `int`.