Established `ControllerController` and action-based views
The `ControllerController` and `*Action` actions are intended to demonstrate and validate the routing and view locations of `/{Area}/{Controller}/{Action}` routes and their ability to fallback to `/Areas/Area/Views/` and `/Views/` locations in order to identify an appropriate action-based view. These will be used by an integration test to make sure that every controller-action location in `TopicViewLocationExpander` is properly called.

This includes wiring up the `ControllerController` via the routes (using the `MapDefaultAreaControllerRoute()` extension method) as well as the `SampleActivator` (by responding to requests for `ControllerController`).