Established `AreaController` and topic-based views
The `AreaController` is intended to demonstrate and validate the routing and view locations of the `/{area}/{controller}/{path}` and `/{area}/{path}` routes and their ability to fallback to `/Areas/Area/Views/` and `/Views/` locations in order to identify an appropriate contenttype-based view. These will be used by an integration test to make sure that every topic-contenttype location in `TopicViewLocationExpander` is properly called.

This includes wiring up the `AreaController` via the routes (using the `MapTopicAreaRoute()` and `MapImplicitAreaControllerRoute()` extension methods) as well as the `SampleActivator` (by responding to requests for `AreaController`).