Established custom `Accordion` action on `AreaController`
One of the business rules of the `TopicViewResultExecutor` is to fallback to the action name for the view, if no other views have been defined. The `Accordion()` action thus allows a test to be created to evaluate this fallback, by specifying that a topic with the content type `ContentList` should use the `Accordion` view (as established in 01b230e2).