Established local `StubTopicRepository` for the integration tests
Database calls are expensive, establish external dependencies, and potentially variable. As such, for the integration tests, we want to use a stub of `ITopicRepository`. We already have a `StubTopicRepository` in the `OnTopic.TestDoubles` project, but it is optimized for the needs of unit tests, not integration tests. This local version of `StubTopicRepository` will serve the same purpose for integration tests, with topics introduced as needed by the integration tests.

This was previously wired up as part of the `SampleActivator` by accident (5fc00636); it should have been committed prior to that. Whoops.