Established container for `TopicViewLocationExpander` tests
The `TopicViewLocationExpander` provides a list of file locations to ASP.NET Core that it should search for a given view. This expands the paths to include the `contentType` as a variable, and looks for paths such as `/Views/ContentTypes/{ContentType}[.{View}].cshtml` and `/Views/{ContentType}/{View}.cshtml`. These tests will be responsible for evaluating the views (f955fa12, 1160903a) and topics (8b38c329,  8e7518ad, 5032a50e) previously established.