Established test for `TopicViewLocationExpander` with topics
The `TopicViewLocationExpander` supports looking for views based on `Topic.ContentType` and an optional view (which can be specified via a number of locations, though we'll use the query string for the purpose of these tests). It evaluates views established as part of a previous commit (1160903a). Each fallback location is given a unique view name so that we don't need to worry about competing views. Each view contains the path to the view so we can confirm the exact view expected was returned.