Established test for `TopicViewLocationExpander` with actions
The `TopicViewLocationExpander` supports looking for views based on `Controller` and an optional `Action`. This expands on the out-of-the-box view locations by allowing established conventions to be prioritized against topic-specific conventions (7049cc94). This test evaluates actions and views established as part of a previous commit (f955fa12). Each fallback location is given a unique view name so that we don't need to worry about competing views. Each view contains the path to the view so we can confirm the exact view expected was returned.