Established container for `TopicViewResultExecutor` tests
The `TopicViewResultExecutor` attempts to determine the optional view name from the query string, `Accept` HTTP headers, controller action name, and `Topic.View` property. If none of those can be identified, then it falls back to the `Topic.ContentType`.

These tests will be responsible for evaluating that the appropriate views (01b230e2) will be loaded based on the appropriate preconditions.