Bug Fix: Views not being determined based on controller action
The view name can be defined from a number of locations, including the identifier of the current controller's action. The action option is not currently working, however, unless a previous option was already evaluated. In practice, this will usually be the case because most requests have an `Accept` header. A common exception to this, however, is for AJAX calls, which may not explicitly include the header. This can easily be fixed by setting the default for the test to `true` if `view.Success` returns null (implying the previous tests haven't been evaluated). This resolves #61.