Establish test for `MapImplicitAreaControllerRoute()`
This is the basic test for `/Area/{action?}/{id?}` or `/Area/{path}/` requests, which will go through `AreaController`, in this particular setup—though it will support any controllers which have the same name as their corresponding area (i.e., in this case, `[Area("Area")]` corresponds to `AreaController`). In this case, the `Controller` needn't be included in the route, as the `MapImplicitAreaControllerRoute()` will implicitly assume `AreaController` based on the `area` route parameter.