Replace current guard clause with `Contract.Requires()`
Typically, `Contract.Requires()` go at the beginning—and this one could be as well. For backward compatibility, we're keeping it where it is. `Contract.Requires()` effectively does the same thing as this method—and resolves CA2208, which expects the exact parameter name to be passed to an `ArgumentNullException`. (We could also use `Contract.Assume()` here, though `Contract.Require()` uses the appropriate `ArgumentNullException` by default.)