Removed empty constructor on `ReadOnlyTopicMultiMap`
The empty constructor was a bit of a hack in order to ensure that the `TopicRelationshipMultiMap` constructor could relay an existing instance of a `TopicMultiMap()` to the base `ReadOnlyTopicMultiMap`. The `ReadOnlyTopicMultiMap` _always_ expects its `Source` property to be initialized, so this relied on the derived class being sure to set this.

A better approach is just to pass a `new TopicRelationshipMultiMap` to the `ReadOnlyTopicMultiMap()` constructor, and then set the local `_storage` backing variable to the `base.Source` property. This effectively does the same thing, but prevents the need for the `protected` empty constructor.

Technically, this is a breaking change. The constructor _should_ have been marked as `internal` since it was only intended for internal use. Fortunately, as this is new with OnTopic 5.0.0, we have high confidence that this hasn't been utilized, and this overload isn't documented in the `README` files, at least, so we feel comfortable removing it.