Allow uninitialized properties after construction
Normally, required `init` only properties should be initialized by the constructor. That's a good rule. For view and binding models, however, we expect these to be initialized not by the constructor, but instead by the e.g. `TopicMappingService` or the `ReverseTopicMappingService`. As such, these should be suppressed. Unfortunately, these can only be suppressed in the source, and not globally via e.g. `[SuppressMessage()]`. That means these need to be applied to each view or binding model with required properties.