Fixed previously updated guard clause for `GetHierarchicalRoot()`
In a previous update, I modified the guard clause for `GetHierarchicalRoot()` to use `Contract.Requires()` (87e42966). But `Contract.Requires(object)` only checks that the object is null, not empty, and thus didn't maintain parity with the previous guard clause which used `IsNullOrEmpty()`.

To remedy this, I've updated the clause to explicitly check `IsNullOrEmpty()`. Since this changes the overload to `Contract.Requires(isValid)`, however, that would result in it throwing an `InvalidOperationException` instead. (This makes logical sense, but is a unintuitive for users of the library.) As such, I'm explicitly setting `ArgumentNullException`.

Finally, since the whole point of using `Contract.Requires()` instead of `Contract.Assumes()` was to take advantage of its implicit use of `ArgumentNullException`, and that advantage doesn't exist with the `Contract.Requires(isValid)` overload, I'm updating this to use `Contract.Assume()` instead, which is more semantically appropriate given the placement.

Eep!