Bug Fix: Return 403 if `PageGroup` has no `IsVisible()` children
A `PageGroup` cannot be displayed; it is just a container of child topics. So when a `PageGroup` is requested, the `[ValidateTopic]` attribute redirects to the first `IsVisible()` child. If no such child exists, a 403 should be returned. But, instead, an exception was being thrown due to the attempt to redirect to a null path. This resolves #62.