Simplify setting of the `contenttype` route parameter
In previous versions of OnTopic, the `contenttype` route parameter was typically set by upstream logic, and thus there was more defensive code to ensure that it was reset in the case that the `TopicController` changed the target `Topic` (as might happen with e.g. a 404). The `contenttype` route parameter is no longer set upstream, however, and so there isn't a need to be as defensive in this case. If it is set, we should be able to trust the value—but, more likely, it won't be.