Update Accept Header test to use multiple values
In a real-world scenario, the HTTP Accept header will contain multiple values, not just one value. As such, the test should reflect this reality. It turns out, this test actually fails—meaning this functionality would _not_ work in most real world scenarios! This issue is captured in #63, and will be fixed in a subsequent commit.