Update Accept Header test to use complex values
While this is a less expected scenario, the HTTP Accept headers can, potentially, include additional metadata with each value such as `level` and `q`. If these are present, they should be stripped out and not evaluated as part of the header. This isn't currently support. This is also being tracked as part of #63, and will be addressed in a subsequent commit.