Bug Fix: Resolved issues with complex HTTP Accept headers
The `TopicViewResultExecutor` will attempt to identify views based on the HTTP Accept headers. This is intended to allow e.g. a `JSON` view to be offered, and requested via the `Accept` headers. The previous code, however, only evaluated the first item, and then failed to account for potential metadata. This update rewrites (and simplifies) the code, thus allowing it to correctly account for these considerations. This resolves #63.