Updated tests not to expect mismatches to be returned with `!isDirty`
Previously, if `IsExtendedAttributeMismatch()` was true, then the attribute would return as both `isDirty` _and_ `!isDirty`. This is unintuitive. If the attribute is _implicitly_ dirty, even if it's explicitly marked as clean, then it should only be returned as dirty, not clean.

This test currently fails as this is the subject of an underlying bug (#44). This will be resolved in subsequent commits.