Updated tests to explicitly check for correct attributes
Instead of _just_ checking the count, _also_ check that the attribute returned is correct. This makes the test more explicit, while also making it easier to debug if there is a mismatch in results (since the failed test will report what the actual value is).

This test currently fails as this is the subject of an underlying bug (#44). This will be resolved in subsequent commits.