Bug Fix: Don't return mismatched attributes on `!IsDirty`
If an attribute is set to `!IsDirty`, but is _implicitly_ dirty due to an `IsExtendedAttributeMismatch()`, then it should not be returned from `GetAttributes()` if the `isDirty` parameter is set to `false`. I.e., `IsExtendedAttributeMismatch()` should trump `AttributeRecord.IsDirty`.

This satisfies the new conditions of the unit test (91bae2aa, 5eca52d5), while also resolving Issue #44.