Updated tests to allow `SetPropertyValue()` to null, empty
Previously, the `MemberDispatch.SetPropertyValue()` didn't support null or empty values outside of strings. If a null or empty value was passed, it was effectively ignored, deferring to whatever the existing value was. This prevents the `MemberDispatcher` from being used to override existing values on properties, which in turn prevents attributes and topic references from being deleted if they are backed by a property. This updates the tests to expect that a null or empty value can be set—and, if the type doesn't support null or empty, then it will default to the types `default` instead.

This isn't yet supported by the `DispatchService`; we're updating the test, and then will fix the underlying issue in a subsequent commit.