Remove the `HasSettableProperty()` check in `SetPropertyValue()`
If this were a `public` API, validating the input would make good sense. In this case, however, as an `internal` API, we already know that `HasSettableProperty()` will have been called before calling `SetPropertyValue()`.

There's a bigger issue, though, as well. `SetPropertyValue()` isn't able to `GetType()` of the `value` if it's `null`, which thus prevents `SetPropertyValue()` from working with `null` values (#43). We could fix this by asking callers to explicitly pass the `type`—but since this can (and will have been) already validated by `HasSettableProperty()`, this isn't preferable.

Instead, I'm taking out the `HasSettableProperty()` validation within `SetPropertyValue()`. This will result in `SetPropertyValue()` throwing e.g. an `InvalidOperationException` if it's used to set the wrong type, or an `ArgumentException` if the type is missing. But that makes sense; if `SetPropertyValue()` is set with incorrect values, we'd _expect_ it to throw an exception, instead of failing silently—or, somewhat silently. (It still returns a `false` in those situations.)