Remove null checks from `SetPropertyValue()`
If a property supports nulls, then `SetPropertyValue()` should be able to set them, as opposed to silently ignoring them. If the property doesn't support nulls, then the underlying reflection APIs will set them to their `default` value instead, which works well for our requirements. This resolves #43, and satisfies the new conditions in the updates to the unit tests (1e2fe680).