Update tests to expect exceptions on unvalidated calls
In previous commits, we removed the `HasSettableProperty()` check from `SetPropertyValue()` (c089322d) and the `HasSettableMethod()` check from `SetMethodValue()` (e3a19906). As a result, setting an invalid reference type or referencing an invalid member will now throw an exception instead of returning `false` and otherwise ignoring the request. Given that, the unit tests for validating this behavior need to be updated to reflect this updated logic.