Update `SetPropertyValue()`, `SetMethodValue()` return type to `void`
Previously, `SetPropertyValue()` and `SetMethodValue()` returned a `bool` indicating a success or failure. This allowed them to fail silently, while allowing the caller to respond accordingly. The preferred technique here is to use `HasSettableProperty()` and `HasSettableMethod()` to detect predictable scenarios ahead of time—and otherwise rely on exceptions to bubble up any errors that still occur.

Given that, the internal validation was removed from `SetPropertyValue()` (b38c2507) and `SetMethodValue()` (e3a19906). Without those, though, there's nothing left to return `false`—and, really, no need to do so anyway. Instead, these should be updated to be `void`, which is a more intuitive and familiar expectation for a setter.

This breaks the unit tests, which this commit fixes, but has no downstream impact on the production classes.

(As an aside, the fact that this change doesn't break any of the production code is a testament to the fact that the `bool` return type wasn't offering any value.)