Introduce test for `SetPropertyValue()` on non-nullable value type
This unit test validates that when a non-nullable value type is set to an empty string, it is instead set to the `default` value. This maps to the the out-of-the-box behavior of the underlying reflection library when working with `null` values, and a useful behavior for our purposes since we treat empty strings as null values for the purposes of the `ITopicMappingService` and `Topic.Attributes` collection.

This test requires introducing (yet another) view model, the `NonNullablePropertyTopicViewModel`, which maps to the similar `NullablePropertyTopicViewModel`, but instead exposes non-nullable types.

This confirms the final piece of Issue #43, and clears the roadblocks for #65 and #42.