Consolidate `SetMethodValue()` overloads
With the removal of the `HasSettableMethod()` validation from the `SetMethodValue()` (e3a19906), there's no longer a need to have separate overloads for setting a string (which will use `GetValueObject()` for converting to value types) or an `object` (which will set directly compatible properties). Indeed, we could have even done this previously, as we had already done with `SetPropertyValue()` (13d43d14). Regardless, it's a piece of overhead and complexity that's even easier to handle now that we don't need to validate the `object` type for references.