Refactored `GetValueObject()` logic for `SetPropertyValue()`
With the consolidation of the `SetMethodValue()` overloads, I refactored how strings were detected, and how `GetValueObject()` was called (1318fdc0). This is a more elegant approach than that which I had previously implemented for the consolidation of `SetPropertyValue()` (13d43d14). Given that, I've applied a similar refactoring to `SetPropertyValue()` for consistency.