Introduced test for enforcing business logic with null values
A bug was identified in which the setting a `TrackedRecord<T>.Value` to `null` would fail if that `TrackedRecord<T>.Key` corresponded to a `[ReferenceSetter]` property on the `Topic`, thus needing to `enforceBusinessLogic` via `PropertyDispatcher.Enforce()` (#65). This was due, fundamentally, to the underlying `MemberDispatch` not supporting nulls (#43), which was fixed in a previous bug fix (6448cc8). This unit test validates that that resolution also resolves #65.