Introduced test for setting `[ReferenceSetter]` properties to null
The bug described in #65, and validated in the previous commit (a4693a0d), was only expected to affect scenarios where `enforceBusinessLogic` caused `PropertyDispatch.Enforce()` to call a `[ReferenceSetter]` via reflection, and thus _shouldn't_ impact direct calls to a `[ReferenceSetter]` via the property. Nevertheless, as this is critical functionality, I'm adding a unit test to validate this behavior continues to work from this end as well, since it's not as simple as just setting a property.