Introduce new empty `Load()` overload for loading the entire graph
Previously, this was handled via the optional `uniqueKey` parameter on the existing `Load()` overload. This made it difficult to validate the `uniqueKey`, however, and made it ambiguous what `Load(null)` meant.

This is potentially pernicious in that an unchecked `uniqueKey` argument could be null—but, in the previous overload, that would be result in the entire topic graph being returned, instead of returning a `null` or `TopicNotFoundException()`.