Implemented empty `Load()` overload on base classes
All `ITopicRepository` implementations are expected to derive from either `ObservableTopicRepository` or `TopicRepositoryDecorator`—directly, or indirectly via e.g. `TopicRepository`. As such, by implementing the new `Load()` overload (72f63f12) on these two base classes, we should make it immediately available on all implementations.

This isn't strictly necessary since the new `Load()` overload is implemented as a default interface method on `ITopicRepository` (72f63f12) and will thus maintain compatibility with anyone referencing an `ITopicRepository` implementation via the interface. And since we expect all interactions to use the interface in dependency injection, that should cover all scenarios. Nevertheless, we also expect the concrete implementations to offer a compatible interface for predictability, so this helps maintain that.

Technically, this is a breaking change in that any derived `ITopicRepository` implementation will need to update the `uniqueKey` parameter on the existing overload to be required, in order to avoid an ambiguous reference. In practice, we have good visibility into the custom `ITopicRepository` implementations, and don't expect any custom implementations to be used outside of the core library. Given that, we feel alright making this change outside the context of a major release.