Validate `uniqueKey` parameter
With the empty `Load()` overload on `ITopicRepository` (72f63f12) and the base classes (20ca77a7), and `uniqueKey` now a required parameter (6f02f4c3), there is no longer any ambiguity between `Load()` and `Load(null)`, and so we can validate the `uniqueKey` parameter.

In theory, this could be seen as a breaking change, since the behavior of `Load(null)` will change. In practice, anyone relying on this is almost certainly getting unexpected results, and likely has an unidentified bug. Anyone _expecting_ the entire topic graph will be calling `Load()`. As such, this should be seen as a bug fix rather than a breaking change.

Finally, the `SqlTopicRepository` will throw an exception if the `uniqueKey` is null or empty, while the others `ITopicRepository` implementations will simply return `null`. This is done for consistency with the existing behavior. We want to reevaluate this in the future—but as that's a major breaking change, it's something we'll need to reevaluate as part of OnTopic 6.0.0. For more information, see #57.